+++ lib/src/config/routes/routes.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/home/account_screen.dart';
-
-import '../../features/daily_news/presentation/pages/article_detail/article_screen.dart';
-import '../../features/daily_news/presentation/pages/discover/discover_screen.dart';
-import '../../features/daily_news/presentation/pages/home/home_screen.dart';
+import 'package:news_app/src/features/daily_news/presentation/article_detail/article_screen.dart';
+import 'package:news_app/src/features/daily_news/presentation/discover/screens/discover_screen.dart';
+import 'package:news_app/src/features/daily_news/presentation/home/screens/account_screen.dart';
+import 'package:news_app/src/features/daily_news/presentation/home/screens/home_screen.dart';
 
 class AppRoutes {
   static Map<String, Widget Function(BuildContext)> routes = {
