+++ lib/src/core/bloc/article/local/local_article_state.dart
@@ -1,7 +1,5 @@
-import 'package:dio/dio.dart';
 import 'package:equatable/equatable.dart';
-
-import '../../../../domain/entities/article.dart';
+import 'package:news_app/src/features/daily_news/domain/entities/article.dart';
 
 abstract class LocalArticlesState extends Equatable {
   final List<ArticleEntity>? articles;
@@ -12,11 +10,10 @@ abstract class LocalArticlesState extends Equatable {
   List<Object> get props => [articles!];
 }
 
-
-class LocalArticlesLoading extends LocalArticlesState{
+class LocalArticlesLoading extends LocalArticlesState {
   const LocalArticlesLoading();
 }
 
-class LocalArticlesDone extends LocalArticlesState{
+class LocalArticlesDone extends LocalArticlesState {
   const LocalArticlesDone(List<ArticleEntity> articles) : super(articles: articles);
 }
