--- lib/src/core/util/date_time_formatter.dart
@@ -1,15 +1,0 @@
-import 'package:intl/intl.dart';
-
-String formatTimeDifference(DateTime dateTime) {
-  Duration difference = DateTime.now().difference(dateTime);
-
-  if (difference.inDays > 0) {
-    return DateFormat.yMMMd().format(dateTime); // Show date if more than a day ago
-  } else if (difference.inHours > 0) {
-    return '${difference.inHours} ${difference.inHours == 1 ? 'hour' : 'hours'} ago';
-  } else if (difference.inMinutes > 0) {
-    return '${difference.inMinutes} ${difference.inMinutes == 1 ? 'minute' : 'minutes'} ago';
-  } else {
-    return 'Just now';
-  }
-}
