--- lib/src/features/daily_news/presentation/bloc/home_navbar/nav_event.dart
@@ -1,9 +1,0 @@
-abstract class NavEvent {
-  const NavEvent();
-}
-
-class NavChangeIndex extends NavEvent {
-  final int newIndex;
-
-  const NavChangeIndex({required this.newIndex});
-}
