--- lib/src/features/daily_news/presentation/bloc/home_navbar/nav_state.dart
@@ -1,14 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class NavState extends Equatable {
-  final int index;
-
-  const NavState({required this.index});
-
-  @override
-  List<Object> get props => [index];
-}
-
-class NavLoaded extends NavState {
-  const NavLoaded(int index) : super(index: index);
-}
