+++ lib/src/features/daily_news/presentation/discover/bloc/discover_event.dart
