+++ lib/src/features/daily_news/presentation/discover/components/category_news.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/discover/bloc/discover_bloc.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/discover/bloc/discover_event.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/discover/bloc/discover_state.dart';
+import 'package:news_app/src/core/components/article/article_tile_small.dart';
+import 'package:news_app/src/features/daily_news/domain/enums/news_category_enum.dart';
 
-import '../../../../../../core/components/article/article_tile_small.dart';
-import '../../../../domain/enums/news_category_enum.dart';
+import '../bloc/discover_bloc.dart';
+import '../bloc/discover_event.dart';
+import '../bloc/discover_state.dart';
 
 class CategoryNews extends StatelessWidget {
   const CategoryNews({
