+++ lib/src/features/daily_news/presentation/discover/screens/discover_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/discover/bloc/discover_bloc.dart';
-import 'package:news_app/src/features/daily_news/presentation/pages/discover/bloc/discover_event.dart';
+import 'package:news_app/src/core/components/bottom_nav_bar.dart';
+import 'package:news_app/src/features/daily_news/domain/enums/news_category_enum.dart';
+import 'package:news_app/src/injection_container.dart';
 
-import '../../../../../core/components/bottom_nav_bar.dart';
-import '../../../../../injection_container.dart';
-import '../../../domain/enums/news_category_enum.dart';
-import 'components/category_news.dart';
-import 'components/discover_news.dart';
+import '../bloc/discover_bloc.dart';
+import '../bloc/discover_event.dart';
+import '../components/category_news.dart';
+import '../components/discover_news.dart';
 
 class DiscoverScreen extends StatelessWidget {
   const DiscoverScreen({super.key});
